Fixed some project and solution generatation errors

+++ scripts/generate.py
@@ -1,24 +1,28 @@
 import os
-from scripts import product, xmler
+from scripts import product, xmler, cmd
 
 def parse(parser):
     parser.add_argument('-p', '--projectPath', type=str, required=True,       help="Path to project")
 
 def createProjectAndSolution(project):
-    csProjFile = "{}/{}.csproj".format(project.name, project.name)
-    csTestProjFile = "{}/{}.csproj".format(project.testName, project.testName)
-    csTypeDProjFile = "{}/{}.csproj".format(project.devModuleName, project.devModuleName)
+    csProjFile = "{}\{}.csproj".format(project.name, project.name)
+    csTestProjFile = "{}\{}.csproj".format(project.testName, project.testName)
+    csTypeDProjFile = "{}\{}.csproj".format(project.devModuleName, project.devModuleName)
     slnFile = "{}.sln".format(project.name)
 
-    os.system("dotnet new sln --name {}".format(project.name))
-    os.system("dotnet new classlib --name {}".format(project.name))
+    if not os.path.exists(slnFile):
+        cmd.exec("dotnet new sln --name {}".format(project.name))
+    if not os.path.exists(csProjFile):
+        cmd.exec("dotnet new classlib --name {}".format(project.name))
     if os.path.exists("{}/Class1.cs".format(project.name)):
         os.remove("{}/Class1.cs".format(project.name))
     if project.haveDevModule:
-        os.system("dotnet new wpflib --name {}".format(project.devModuleName))
+        if not os.path.exists(csTypeDProjFile):
+            cmd.exec("dotnet new wpflib --name {}".format(project.devModuleName))
         if os.path.exists("{}/Class1.cs".format(project.devModuleName)):
             os.remove("{}/Class1.cs".format(project.devModuleName))
-        os.system("dotnet add {} reference {}".format(csTypeDProjFile, csProjFile))
+        if not cmd.exist("dotnet list {} reference".format(csTypeDProjFile), csProjFile):
+            cmd.exec("dotnet add {} reference {}".format(csTypeDProjFile, csProjFile))
     
     if project.haveTest:
         testFile = "{}/UnitTest1.cs".format(project.testName)
@@ -31,26 +35,34 @@ def createProjectAndSolution(project):
         if os.path.exists(testUsingFile):
             testUsingFileExisted = True
             os.rename(testUsingFile, testUsingFileTemp)
-        os.system("dotnet new xunit --name {}".format(project.testName))
-        os.system("dotnet add {} reference {}".format(csTestProjFile, csProjFile))
+        if not os.path.exists(csTestProjFile):
+            cmd.exec("dotnet new xunit --name {}".format(project.testName))
+        if not cmd.exist("dotnet list {} reference".format(csTestProjFile), csProjFile):
+            cmd.exec("dotnet add {} reference {}".format(csTestProjFile, csProjFile))
         if project.haveDevModule:
             csProjTestXML = xmler.load(csTestProjFile)
             propertyGroup = xmler.getElementWithName(csProjTestXML, "PropertyGroup")
             xmler.setElementWithValue(propertyGroup, "TargetFramework", "{}-windows".format(xmler.getElementWithName(propertyGroup, "TargetFramework").firstChild.data))
             xmler.getOrCreateElementWithValue(csProjTestXML, propertyGroup, "UseWPF", "true")
             xmler.save(csProjTestXML, csTestProjFile)
-            os.system("dotnet add {} reference {}".format(csTestProjFile, csTypeDProjFile))
-        os.remove(testUsingFile)
+            if not cmd.exist("dotnet list {} reference".format(csTestProjFile), csTypeDProjFile):
+                cmd.exec("dotnet add {} reference {}".format(csTestProjFile, csTypeDProjFile))
+        if(os.path.exists(testUsingFile)):
+            os.remove(testUsingFile)
         if testUsingFileExisted:
             os.rename(testUsingFileTemp, testUsingFile)
         if not testFileFound:
-            os.remove(testFile)
+            if(os.path.exists(testFile)):
+                os.remove(testFile)
 
-    os.system("dotnet sln {} add {}".format(slnFile, csProjFile))
+    if not cmd.exist("dotnet sln {} list".format(slnFile), csProjFile):
+        cmd.exec("dotnet sln {} add {}".format(slnFile, csProjFile))
     if project.haveDevModule:
-        os.system("dotnet sln {} add {}".format(slnFile, csTypeDProjFile))
+        if not cmd.exist("dotnet sln {} list".format(slnFile), csTypeDProjFile):
+            cmd.exec("dotnet sln {} add {}".format(slnFile, csTypeDProjFile))
     if project.haveTest:
-        os.system("dotnet sln {} add {}".format(slnFile, csTestProjFile))
+        if not cmd.exist("dotnet sln {} list".format(slnFile), csTestProjFile):
+            cmd.exec("dotnet sln {} add {}".format(slnFile, csTestProjFile))
 
 def addPreBuildEvents(csProjXML):
     target = xmler.getOrCreateElementWithAttributes(csProjXML, csProjXML, "Target",[
